refactor(infra/bedrock): commonをbaseとpromptsに再構成（TDDで開発）
- BaseBedrockClientを bedrock/base/BedrockClient.ts へ移動
- プロンプト系ユーティリティを src/shared/ai/prompts/utils.ts に分離
- 既存のimportを新配置に置換（後方互換レイヤは設置せず）
- Red→Green: 新パス向けユニットテスト追加し、実装でグリーン化

Refs: #9
