+++ src/shared/ai/prompts/utils.ts
