+++ src/shared/infrastructure/external/bedrock/base/BedrockClient.ts
