--- src/shared/infrastructure/external/bedrock/common/index.ts
@@ -1,2 +1,0 @@
-export { BaseBedrockClient } from "./baseClient";
-export * from "./middleware";
